requirements.txt
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
appdirs==1.4.4
arrow==1.2.3
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
audioread==3.0.0
av==10.0.0
beartype==0.12.0
beautifulsoup4==4.12.2
blessed==1.20.0
braceexpand==0.1.7
cachetools==5.3.0
click==8.1.3
contourpy==1.0.7
croniter==1.3.10
cycler==0.11.0
dataclasses-json==0.5.8
dateutils==0.6.12
decord==0.6.0
deepdiff==6.3.0
dtk==0.2
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.88.0
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.0
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gammatone==1.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
hickle==5.0.2
huggingface-hub==0.14.1
humanize==4.6.0
imageio==2.27.0
inquirer==3.1.3
ipdb==0.13.13
itsdangerous==2.1.2
jedi==0.18.2
jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.216
langchainplus-sdk==0.0.17
lazy-loader==0.2
librosa==0.10.0.post2
lightning==2.0.0
lightning-cloud==0.5.33
lightning-utilities==0.8.0
llvmlite==0.39.1
markdown==3.4.3
markdown-it-py==2.2.0
markupsafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mdurl==0.1.2
mergedeep==1.3.4
mock==5.0.2
msgpack==1.0.5
msgpack-numpy==0.4.8
multidict==6.0.4
musdb==0.4.0
mypy-extensions==1.0.0
networkx==3.1
nose==1.3.7
numba==0.56.4
numexpr==2.8.4
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
ordered-set==4.1.0
outcome==1.2.0
pandas==1.5.3
panns-inference==0.1.0
pesq==0.0.4
pillow==9.5.0
pooch==1.6.0
prompt-toolkit==3.0.38
protobuf==4.22.3
pyaml==23.5.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pygments==2.14.0
pyjwt==2.6.0
pyloudnorm==0.1.1
pyparsing==3.0.9
pystoi==0.3.3
python-editor==1.0.4
python-multipart==0.0.6
pytorch-ignite==0.3.0
pytorch-lightning==2.0.1.post0
pytz==2023.3
pywavelets==1.4.1
pyyaml==6.0
readchar==4.0.5
regex==2023.3.23
requests-oauthlib==1.3.1
resampy==0.4.2
rich==13.3.3
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
selenium==4.8.3
simplejpeg==1.6.6
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4
soxr==0.3.5
sqlalchemy==2.0.17
stack-data==0.6.2
starlette==0.22.0
starsessions==1.3.0
stempeg==0.2.3
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2023.3.21
timm==0.3.2
tokenizers==0.13.3
tomli==2.0.1
torchfile==0.1.0
torchlibrosa==0.1.0
torchmetrics==0.11.4
traitlets==5.9.0
transformers==4.28.1
trio==0.22.0
trio-websocket==0.10.2
typeguard==3.0.2
typing-extensions==4.5.0
typing-inspect==0.9.0
uvicorn==0.21.1
visdom==0.1.8.9
wcwidth==0.2.6
webdataset==0.2.48
websocket-client==1.5.1
websockets==11.0.1
werkzeug==2.2.3
wget==3.2
wsproto==1.2.0
yarl==1.8.2
zenodo-get==1.3.4
zsvision==0.7.8